requirements.txt
typer==0.7.0
rich==13.3.3
aiohttp==3.8.4
aiofiles==23.1.0
pyyaml
requests
pydantic
GitPython
